ccViewportParameters and ccGenericGLDisplay split in two files
